space-leap: fix WebGPL image loading for Chrome for star library
The WebGL loading of star stuff does not work on Chrome but does
on Firefox, the issue Chrome complains about is:

Uncaught SecurityError: Failed to execute 'texImage2D' on
'WebGLRenderingContext': the cross-origin image at img/star.png may not be loaded.

The solution to this is to add an image.onload() with the respective loader.
This adds this solution to the library.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>