space-leap: use THREE.ImageUtils.crossOrgin
The previous commit addressed the Chrome complaint on
cross-origin image usage but using native APIs. Since
we are using a library for image loading we also need
to make sure to propagate the crossOrgin = "anonymous"
tag to the THREE.ImageUtils.

If you keep running into this issue consider the security
risk of having a web page refer to a local image on your
hard drive. This seems to therefrore be a cross domain
security fix -- not a bug. To resolve this the best
advice comes from Robbie Tilton [0] [1]:

================================================================
If you're new to three.js and you're using Chrome to run
the web app locally you'll receive cross-domain issues with
loading in the files.  To run it locally - there is a simple
python command you can use in your terminal:

	python -m SimpleHTTPServer 8000

This will resolve cross-domain issues with Chrome and the
example will be viewable on localhost:8000
================================================================

[0] http://www.robbietilton.com/
[1] http://codepen.io/leapmotion/pen/cEilz

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>