space-leap: import base earth ground surface map
This comes from:

http://commons.wikimedia.org/wiki/File:Srtm_ramp2.world.21600x10800.jpg

This image is licensed under the Public Domain. As per inspiration from
Robbie's work we are going to use the same surface map for earth before
we jump to asteroids. Robbie's work again:

http://codepen.io/leapmotion/pen/cEilz

Interested parties in looking for more maps of earth maps can use
NASA's earthobservatory image database:

http://earthobservatory.nasa.gov/IOTD/view.php?id=42805

As for clouds the image we import is a 3600x1800 resolution image from
NASA's cloud fraction page:

http://neo.sci.gsfc.nasa.gov/view.php?datasetId=MODAL2_M_CLD_FR

The cloud image we took is the view of the clouds as of August 2013,
which I am not sure why its that date given its April 2014. I tried
my best to look for licensing information about the cloud image
but I cannot find any, so I am in hopes NASA is OK to release that
under the Public Domain. I'll poke and verify.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>